requirements.txt
python-telegram-bot==21.6
anthropic>=0.40.0
openai==1.51.0
python-dotenv==1.0.1
apscheduler==3.10.4
aiosqlite==0.20.0
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic>=2.0
